fetchdata change
